move wind to simulation
